Refactor: ticket controller, service 로직 분리
